+++ app/src/app/api/token-price/route.ts
@@ -1,6 +1,6 @@
 import { tokenPriceSchema } from '@/models/api-schemas'
-import { getCoingekoId } from '@/models/token'
 import { CACHE_REVALIDATE_IN_SECONDS, getTokenPrice } from '@/services/balance'
+import { getCoingekoId } from '@/utils/token'
 import { getErrorMessage } from '@/utils/transferTracking'
 import { unstable_cache } from 'next/cache'
 import { NextResponse } from 'next/server'
