+++ app/src/models/token.ts
@@ -1,7 +1,4 @@
-import { REGISTRY } from '@/registry/mainnet/mainnet'
 import { TMultiLocation } from '@paraspell/sdk'
-import { deepEqual } from '@paraspell/sdk-common'
-import { captureException } from '@sentry/nextjs'
 
 export interface Token {
   id: string
@@ -28,20 +25,3 @@ export type Bridge = 'Snowbridge' | 'Other'
 
 // The origin of a token
 export type Origin = { type: 'Ethereum'; bridge: Bridge } | { type: 'Polkadot'; paraId: number }
-
-export function getCoingekoId(token: Token): string {
-  return token.coingeckoId ?? token.name.toLocaleLowerCase().replaceAll(' ', '-')
-}
-
-export function getTokenByMultilocation(multilocation: TMultiLocation): Token | undefined {
-  const token = REGISTRY.tokens.find(token => deepEqual(token.multilocation, multilocation))
-
-  // Usually a token should be found, so log a warning if not
-  if (!token)
-    captureException(new Error('Token not found by multilocation'), {
-      level: 'warning',
-      extra: { multilocation },
-    })
-
-  return token
-}
